// src/Components/Forms/FormAssistance/types.ts
export type FormValues = {
    midWeek1: number | undefined
    midWeek2: number | undefined
    midWeek3: number | undefined
    midWeek4: number | undefined
    midWeek5: number | undefined
    endWeek1: number | undefined
    endWeek2: number | undefined
    endWeek3: number | undefined
    endWeek4: number | undefined
    endWeek5: number | undefined
}